feat: verify inputs
